requirements.txt
requests~=2.28.2
selenium~=4.18.1
jira~=3.6.0
python-dotenv~=1.0.1
utils~=1.0.2